+++ database/migrations/2025_07_22_041417_create_random_consortia_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('random_consortia', function (Blueprint $table) {
-            $table->id();
-            $table->string('title')->nullable();
-            $table->unsignedBigInteger('language_id');
-            $table->foreign('language_id')->references('id')->on('languages')->onDelete('cascade');
-            $table->enum('style', ['style1'])->default('style1');
-            $table->text('description')->nullable();
-            $table->text('short_description')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('random_consortia')) {
+            Schema::create('random_consortia', function (Blueprint $table) {
+                $table->id();
+                $table->string('title')->nullable();
+                $table->unsignedBigInteger('language_id');
+                $table->foreign('language_id')->references('id')->on('languages')->onDelete('cascade');
+                $table->enum('style', ['style1'])->default('style1');
+                $table->text('description')->nullable();
+                $table->text('short_description')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
